Don't auto-install DependencyPatch immediately
Instead, the code is wired into `Bundler` and `Gem` only after all code
of `Bundler::Override` has been sourced.
