+++ lib/bundler-override.rb
@@ -31,6 +31,22 @@ end
 require_relative "bundler/override/dependency_patch"
 require_relative "bundler/override/dsl_patch"
 
+module Bundler
+  class RemoteSpecification
+    include Override::DependencyPatch
+  end
+
+  class EndpointSpecification
+    include Override::DependencyPatch
+  end
+end
+
+module Gem
+  class Specification
+    include Bundler::Override::DependencyPatch
+  end
+end
+
 Bundler::Dsl.prepend(Bundler::Override::DslPatch)
 ObjectSpace.each_object(Bundler::Dsl) do |o|
   o.singleton_class.prepend(Bundler::Override::DslPatch)
